models-prepration/motion-encoder/requirements.txt
# Core ML and Deep Learning
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Data Processing and Analysis
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# Configuration Management
PyYAML>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress Bars and Utilities
tqdm>=4.62.0

# Logging and Monitoring
tensorboard>=2.8.0
wandb>=0.12.0  # Optional: Weights & Biases for experiment tracking

# File I/O and Serialization
pickle-mixin>=1.0.2
joblib>=1.1.0

# Mathematical Operations
numba>=0.56.0  # Optional: For performance optimization

# Development and Testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Type Hints
typing-extensions>=4.0.0

# Memory Profiling (Optional)
memory-profiler>=0.60.0
psutil>=5.8.0

# GPU Monitoring (Optional)
nvidia-ml-py3>=7.352.0

# Additional Utilities
click>=8.0.0  # For command line interfaces
rich>=12.0.0  # For rich terminal output
pathlib2>=2.3.0  # Enhanced path handling

# Data Validation
cerberus>=1.3.0

# Performance Monitoring
line-profiler>=3.5.0

# Jupyter Notebook Support (Optional)
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0